멘토님 상담이후 수정1
