+++ src/components/Login/LOGIN_DATA/FIRSTWINDOW_DATA.js
@@ -1,0 +1,5 @@
+const FIRSTWINDOW_DATA = [
+  { infoType: 'email', inputType: 'email', string: '이메일 주소' },
+];
+
+export default FIRSTWINDOW_DATA;
