+++ src/components/Login/LOGIN_DATA/RESETPW_DATA.js
@@ -1,0 +1,5 @@
+const RESETPW_DATA = [
+  { infoType: 'email', type: 'email', text: '이메일 주소' },
+];
+
+export default RESETPW_DATA;
