+++ src/components/Login/LOGIN_DATA/SIGNIN_DATA.js
@@ -1,0 +1,6 @@
+const SIGNIN_DATA = [
+  { infoType: 'email', inputType: 'email', string: '이메일 주소' },
+  { infoType: 'password', inputType: 'password', string: '패스워드' },
+];
+
+export default SIGNIN_DATA;
