+++ src/components/Login/LOGIN_DATA/SIGNUP_DATA.js
@@ -1,0 +1,9 @@
+const SIGNUP_DATA = [
+  { infoType: 'email', inputType: 'email', string: '이메일 주소' },
+  { infoType: 'password', inputType: 'password', string: '패스워드' },
+  { infoType: 'rePassword', inputType: 'password', string: '패스워드 확인' },
+  { infoType: 'lastName', inputType: 'text', string: '성' },
+  { infoType: 'firstName', inputType: 'text', string: '이름' },
+];
+
+export default SIGNUP_DATA;
