+++ src/pages/Login/Login.js
@@ -1,6 +1,10 @@
 import React, { useState } from 'react';
 import LoginCard from '../../components/Login/LoginCard';
 import LoginForm from './LoginForm';
+import FIRSTWINDOW_DATA from '../../components/Login/LOGIN_DATA/FIRSTWINDOW_DATA';
+import RESETPW_DATA from '../../components/Login/LOGIN_DATA/RESETPW_DATA';
+import SIGNIN_DATA from '../../components/Login/LOGIN_DATA/SIGNIN_DATA';
+import SIGNUP_DATA from '../../components/Login/LOGIN_DATA/SIGNUP_DATA';
 import './Login.scss';
 
 const Login = () => {
@@ -25,27 +29,6 @@ const Login = () => {
       />
     );
   };
-  const FIRSTWINDOW_DATA = [
-    { infoType: 'email', inputType: 'email', string: '이메일 주소' },
-  ];
-
-  const SIGNIN_DATA = [
-    { infoType: 'email', inputType: 'email', string: '이메일 주소' },
-    { infoType: 'password', inputType: 'password', string: '패스워드' },
-  ];
-
-  const SIGNUP_DATA = [
-    { infoType: 'email', inputType: 'email', string: '이메일 주소' },
-    { infoType: 'password', inputType: 'password', string: '패스워드' },
-    { infoType: 'rePassword', inputType: 'password', string: '패스워드 확인' },
-    { infoType: 'lastName', inputType: 'text', string: '성' },
-    { infoType: 'firstName', inputType: 'text', string: '이름' },
-  ];
-
-  const RESETPW_DATA = [
-    { infoType: 'email', type: 'email', text: '이메일 주소' },
-  ];
-
   return (
     <div className="login">
       <div
